Refactor useRecentTransactions hook
Refactor the useRecentTransactions hook into smaller files for better organization and maintainability.
