+++ src/hooks/types/useRecentTransactionsTypes.ts
@@ -1,0 +1,6 @@
+
+export interface UseRecentTransactionsProps {
+  projectId?: string | null;
+  dateRange?: { from: Date | null, to: Date | null } | null;
+  managerId?: string | null;
+}
